+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add debug log when a function query fails https://github.com/heroku/cutlass/pull/17
+
 ## 0.2.1
 
 - Fix incorrect conversion of a ProcessStatus into an exit code https://github.com/heroku/cutlass/pull/16